refactor: Cleanup in storage
